refactor and move
